requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
psycopg2-binary>=2.9.0
httpx>=0.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
python-multipart>=0.0.6
